+++ jans-keycloak-link/server/src/main/java/io/jans/keycloak/link/timer/JansKeycloakLinkTimer.java
@@ -764,6 +764,7 @@ public class JansKeycloakLinkTimer extends BaseJansLinkTimer {
 					jansSimplePerson.getCustomAttributes().add(jansCustomAttribute);
 				}
 				jansSimplePerson.setDn(personService.getDnForPerson(personService.generateInumForNewPersonImpl()));
+				jansSimplePerson.setSourceServerName("Keycloak to jans");
 
 				sourceJansSimplePerson.add(jansSimplePerson);
